Adding CourseUserGroupMembership
The purpose of this new model relation class is to prevent race conditions
which can lead to a user being in multiple cohorts in a single course.

The way I've elected to handle this is to use a "version" field on the
Membership model. By using a filter->update approach to change values
stored in the database, we've effectively implmented optimistic locking.
Further reading: http://stackoverflow.com/questions/320096/django-how-can-i-protect-against-concurrent-modification-of-database-entries

This has been tested to the point of confirming that it handles the simple
use cases properly - users can be added to cohorts and will be removed from
any other cohort they may be in as a part of this operation. I still need
to test (using sleep() calls to make any exploits able to bit hit more easily)
and confirm the functionality of the race-case blocking code.

There are a few open questions as well, this PR is far from ready to commit.